Issue 12, move parte do django (backend) para uma pasta separada, assim na raiz fica apenas coisas fora do backend.
