+++ template/setup.cfg
@@ -1,5 +1,5 @@
 [tool:pytest]
-DJANGO_SETTINGS_MODULE = {{ name }}.settings
+DJANGO_SETTINGS_MODULE = {{name}}.{{ name }}.settings
 python_files = tests.py test_*.py *_tests.py
 addopts = -x -vv
