+++ template/uwsgi.ini
@@ -1,8 +1,8 @@
 [uwsgi]
 
-env = DJANGO_SETTINGS_MODULE={{name}}.settings
+env = DJANGO_SETTINGS_MODULE={{name}}.{{name}}.settings
 chdir           = /app
-module          = {{name}}.wsgi
+module          = {{name}}.{{name}}.wsgi
 master          = true
 processes       = 3
 http            = :8000
