+++ template/{{name}}/commons/django_views_utils.py
