+++ template/{{name}}/core/models.py
