+++ template/{{name}}/core/urls.py
