+++ template/{{name}}/{{name}}/urls.py
@@ -19,5 +19,5 @@ from django.urls import path, include
 urlpatterns = [
     path('admin/', admin.site.urls),
     # path('explorer/', include('explorer.urls')),
-    path('api/', include('core.urls')),
+    path('api/', include('{{name}}.core.urls')),
 ]
